added ratio

+++ code/sequential_inference_allparams.py
@@ -182,7 +182,7 @@ def main(argv):
 
         print(i, j)
 
-        num_sim_round = num_sim * (start_num / 10)
+        num_sim_round = int(num_sim * (start_num / 10))
 
         print('number of simulations in this round', num_sim_round)
 
@@ -249,7 +249,7 @@ def main(argv):
 
     start_time = datetime.datetime.now()
 
-    num_sim_round = num_sim * (start_num / 10)
+    num_sim_round = int(num_sim * (start_num / 10))
 
     theta, x_without = inference.run_sim_theta_x(
         prior_i,
